requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
apscheduler
python-telegram-bot
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
requests
beautifulsoup4
yagmail
pydantic
